Working on the cart functionality. Made a CartContent component. Moved handleQuantityChange to utils.
