+++ src/utils/index.js
@@ -2,5 +2,6 @@ import { filterProducts } from "./filterProducts";
 import { calculateDiscount } from "./calculateDiscount";
 import { groupProductsById } from "./groupProductsById";
 import { addToCart } from "./addToCart";
+import { handleQuantityChange } from "./handleQuantityChange";
 
-export { filterProducts, calculateDiscount, groupProductsById, addToCart };
+export { filterProducts, calculateDiscount, groupProductsById, addToCart, handleQuantityChange };
